+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased changes
 
+- `update_form/4` on conn behaves like live views, where unreferenced inputs are unset.
 - Conn driver form functions handle select inputs.
 
 ## 1.2.0