+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased changes
 
+## 1.3.0
+
 - `update_form/4` on conn behaves like live views, where unreferenced inputs are unset.
 - Conn driver form functions handle select, radio, and checkbox inputs.